Task-item-teams-and-dashboards

+++ src/app/components/task-list/task-list.component.ts
@@ -2,24 +2,34 @@ import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatDialog, MatDialogModule } from '@angular/material/dialog';
 import { HttpClient } from '@angular/common/http';
-import { TaskService, Task } from '../../task.service';
+import { CdkDragDrop, DragDropModule, moveItemInArray, transferArrayItem } from '@angular/cdk/drag-drop';
+import { TaskService, Task, Team } from '../../task.service';
 import { StatusManagerComponent } from '../status-manager/status-manager.component';
+import { TaskItemComponent } from '../task-item/task-item.component';
+import { TeamComponent } from '../team/team.component';
 import { StatusService, Status } from '../../status.service';
+import { FormsModule } from '@angular/forms';
+import { RouterLink } from '@angular/router';
 
 @Component({
   selector: 'app-task-list',
   standalone: true,
-  imports: [CommonModule, MatDialogModule],
+  imports: [CommonModule, MatDialogModule, DragDropModule, FormsModule, RouterLink],
   templateUrl: './task-list.component.html',
   styleUrls: ['./task-list.component.css']
 })
 export class TaskListComponent implements OnInit {
   isMenuOpen = true;
   tasks: Task[] = [];
+  teams: Team[] = [];
+  tasksWithTeams: { task: Task, team: Team }[] = [];
   statuses: Status[] = [];
   hoveredDescription: string | null = null;
   activeStatusMenu: string | null = null;
-  newTaskTitle = '';
+  newTaskTitles: { [statusTitle: string]: string } = {};
+  draggingStatusId: string | null = null;
+  searchTerm: string = ''; // Armazena o termo de busca
+  filteredTasks: Task[] = []; // Armazena as tarefas filtradas
 
   constructor(
     private taskService: TaskService,
@@ -31,13 +41,41 @@ export class TaskListComponent implements OnInit {
   ngOnInit(): void {
     this.loadStatuses();
     this.loadTasks();
+    this.loadTasksAndTeams();
+  }
+  
+  loadTasksAndTeams(): void {
+    this.taskService.getTasks().subscribe(tasks => {
+      this.tasks = tasks;
+      this.taskService.getTeams().subscribe(teams => {
+        this.teams = teams;
+        this.combineTasksWithTeams();
+      });
+    });
+  }
+
+  combineTasksWithTeams(): void {
+    this.tasks.forEach(task => {
+      task.team = this.teams.find(t => t.id === task.teamId);
+    });
+  }
+
+  getTeamMemberName(task: Task): string {
+    return task.team ? task.team.name : 'Ownerless';
   }
 
   toggleMenu() {
     this.isMenuOpen = !this.isMenuOpen;
   }
 
-  openModal(event: MouseEvent) {
+  openModalTeam(event: MouseEvent) {
+    event.preventDefault();
+    this.dialog.open(TeamComponent, {
+      width: '500px',
+    });
+  }
+
+  openModalStatus(event: MouseEvent) {
     event.preventDefault();
     this.dialog.open(StatusManagerComponent, {
       width: '500px',
@@ -45,9 +83,18 @@ export class TaskListComponent implements OnInit {
     });
   }
 
+  openModalTaskItem(event: MouseEvent, taskId: string) {
+    event.preventDefault();
+    this.dialog.open(TaskItemComponent, {
+      width: '500px',
+      data: {taskId}
+    });
+  }
+
+
   loadStatuses() {
     this.statusService.getStatuses().subscribe(statuses => {
-      this.statuses = statuses;
+      this.statuses = statuses.sort((a, b) => a.order - b.order);
     });
   }
 
@@ -56,19 +103,29 @@ export class TaskListComponent implements OnInit {
       width: '500px',
       data: { status }
     });
-  
+
     dialogRef.afterClosed().subscribe(result => {
       if (result) {
         this.loadStatuses();
+        this.loadTasks();
       }
     });
   }
 
   deleteStatus(status: Status) {
     console.log('Deleting status:', status.title);
-    this.statusService.deleteStatus(status.id).subscribe(() => {
-      this.statuses = this.statuses.filter(s => s.id !== status.id);
-    });
+    const tasksForStatus = this.getTasksByStatus(status.title);
+    if (tasksForStatus.length > 0) {
+      alert('This status cannot be deleted because there are tasks currently associated with it.');
+      return; 
+  }
+
+    const confirmation = confirm('Are you sure you want to delete this status?');
+    if (confirmation) {
+      this.statusService.deleteStatus(status.id).subscribe(() => {
+        this.statuses = this.statuses.filter(s => s.id !== status.id);
+      });
+    }
   }
 
   showDescription(description: string) {
@@ -93,15 +150,22 @@ export class TaskListComponent implements OnInit {
     return this.tasks.filter(task => task.status === statusTitle);
   }
 
-  addTask(title: string): void {
+  addTask(title: string, status: Status): void {
     if (!title.trim()) {
       return;
     }
 
-    const newTask: Task = { id: 0, title, status: 'unstarted' };
-    this.taskService.addTask(newTask).subscribe(task => {
-      this.tasks.push(task);
-      this.newTaskTitle = '';
+    const newTask: Omit<Task, 'id'> = {
+      title, status: status.title,
+      teamId: '',
+      team: undefined,
+      situation: '',
+      description: '',
+      statusColor: ''
+    };
+  this.taskService.addTask(newTask).subscribe(task => {
+    this.tasks.push(task);
+    this.newTaskTitles[status.title] = '';
     });
   }
 
@@ -116,8 +180,81 @@ export class TaskListComponent implements OnInit {
   }
 
   deleteTask(task: Task): void {
-    this.taskService.deleteTask(task.id).subscribe(() => {
-      this.tasks = this.tasks.filter(t => t.id !== task.id);
+    const confirmation = confirm('Are you sure you want to delete this task?');
+    if (confirmation) {
+      this.taskService.deleteTask(task.id).subscribe(() => { 
+        this.tasks = this.tasks.filter(t => t.id !== task.id);
+      });
+    }
+  }
+
+  onDropTask(event: CdkDragDrop<Task[]>, status: Status): void {
+    const task = event.item.data as Task;
+
+    if (event.previousContainer === event.container) {
+
+      moveItemInArray(event.container.data, event.previousIndex, event.currentIndex);
+    } else {
+
+      task.status = status.title;
+
+      this.taskService.updateTask(task).subscribe(() => {
+
+        transferArrayItem(
+          event.previousContainer.data,
+          event.container.data,
+          event.previousIndex,
+          event.currentIndex
+        );
+      });
+    }
+  }
+
+  onSearch(): void {
+    if (this.searchTerm.trim()) {
+      this.filteredTasks = this.tasks.filter(task =>
+        task.title.toLowerCase().includes(this.searchTerm.toLowerCase())
+      );
+    } else {
+      this.filteredTasks = [];
+    }
+  }
+
+  openTaskItem(taskId: string): void {
+    this.dialog.open(TaskItemComponent, {
+      width: '500px',
+      data: { taskId }
+    });
+  }
+  
+  onDropStatus(event: CdkDragDrop<Status[]>): void {
+  
+    if (event.previousIndex === event.currentIndex) {
+      return;
+    }
+  
+    moveItemInArray(this.statuses, event.previousIndex, event.currentIndex);
+  
+    this.statuses.forEach((status, index) => {
+      status.order = index;
+    });
+  
+    this.updateStatusOrder();
+  }
+
+  updateStatusOrder() {
+
+    this.statuses.forEach((status) => {
+      const updatedStatus = { ...status };
+  
+      this.http.put(`http://localhost:3000/status/${status.id}`, updatedStatus).subscribe(
+        () => {
+          console.log(`Status ID ${status.id} updated successfully`);
+        },
+        (error) => {
+          console.error('Error updating status order:', error);
+        }
+      );
     });
   }
 }
